The hopper would not parse additional leads if a 2nd, 3rd, 4th, 5th, or 6th NEW option was selected and there was not an equal or greater number of non-NEW leads to mix properly.
I modified the hopper program to add the queued NEW leads to the hopper if the NEW count method was in use and their were not enough leads to satisfy the hopper quota.  I increased the NEW_level (which is hopper_level/NEW_count) to equal the hopper_level (or NEW_lead_count, whichever is greater) should there be enough dialable NEW to do so.  The actual count of pulled NEW leads is withheld until after the hopper process has completed for the campaign, they withheld count of NEW leads is then added to the end of the hopper.