Added an "Extensions" panel to admin. Create a simple dialplan builder for manual extensions.
Redesigned the Campaign IVR so that it can be used without the Virtual
Agents, then branded it under Extensions as Extension IVR. It doesn't do
any call tracking, it is just simple menus that can be built and can be attached
to any "Extension".

Modified the osdial_campaigns table to reference the ivr_id of
osdial_ivr and removed the campaign_id fieild. This allows the IVR
components to be reused for for Extension IVR.

Modified agi-OSDivr.agi to bypass lead status and log updates when in
"Extension" mode.

Loaded static extension entries listed in osdial_astgen.pl into the
osdial_extensions tables. Added routine to load all extensions from
osdial_extensions.

Modified "Extensions" drop-down to be aware of the new osdial_extensions
table.

Updated db server version to 3.0.1.105.