Summary: Changed how common commands like muting and volume adjustments are processed. An Originate would be sent and connected to an extension that would run the maintainance on the the conference. Now we are using AMI events directly for these, if not the "Command" event, the "Originate" is still used in many cases for commands, but instead of it of it calling into the dialplan, we pass it the fields Application and Data, and it calls the function directly.
Worked on lookback IAX accounts. Changed default IAX options.

Removed all of the dialplan extensions in the database, flat-files, and
generators for the functions that were replaced with the direct Command
and Application calls. Added a small loop to declare new contexts as
added to the Extensions.

Changed the MonitorConf routine to send the filename as a channel
vairable (FILENAME) instead of using the callerid. The callerid was limited to 40
characters, where channel vars are not.

Bumped to version 3.0.1.119- Made minor changes to the database,
increased all of the cmd_line fields in osdial_manager to 1024 chars, as
some were getting clipped at 100.  Updated the databases entries for the
recording extensions, 8309, 8310, 8311, to use the FILENAME channel variable.